+++ events.py
@@ -47,21 +47,31 @@ class EventsHandler(base.BaseHandler):
 
 
 class TimeHandler(base.BaseHandler):
+    @staticmethod
+    def parse_available_times(value):
+        try:
+            tokens = unquote(value).split(',')
+            if any(not token for token in tokens):
+                raise ValueError
+            return [int(token) for token in tokens]
+        except (TypeError, ValueError):
+            raise tornado.web.HTTPError(400)
+
     @tornado.web.authenticated
     async def post(self):
         _user_id = self.get_current_user()['id']
         _user_name = self.get_current_user()['name']
         _event_id = self.get_int_argument('event_id')
         await self.require_event_access(_event_id)
-        _times = self.get_argument('availabletimes')
+        _times = self.parse_available_times(self.get_argument('availabletimes'))
         self.db.execute(
             "DELETE FROM willbeout_availability WHERE event_id = %s and user_id = %s",
             _event_id, int(_user_id))
 
-        for i in unquote(_times).split(','):
+        for i in _times:
             self.db.execute(
                 """INSERT INTO willbeout_availability (user_id, user_name, time, event_id) VALUES (%s, %s, %s, %s)""",
-                int(_user_id), _user_name, int(i), _event_id)
+                int(_user_id), _user_name, i, _event_id)
         self.redirect('/event?event_id=' + str(_event_id))
 
     @tornado.web.authenticated